refactor(gui): split up Timeline widget into multiple
Also moves composables to different files in a new subfolder.
This should improve readability of the code.
